Reshuffle things around for better CMSSW compatibility
