+++ DiscretePF2Firmware.h
@@ -1,6 +1,9 @@
+#ifndef FASTPUPPI_NTUPLERPRODUCER_DISCRETEPF2FIRMWARE_H
+#define FASTPUPPI_NTUPLERPRODUCER_DISCRETEPF2FIRMWARE_H
 
-#include "DiscretePFInputs.h"
-#include "src/data.h"
+/// NOTE: this include is not standalone, since the path to DiscretePFInputs is different in CMSSW & Vivado_HLS
+
+#include "firmware/data.h"
 #include <vector>
 
 namespace dpf2fw {
@@ -41,4 +44,6 @@ namespace dpf2fw {
     }
 
 
-};
+} // namespace
+
+#endif
