+++ pattern_serializer.h
@@ -1,4 +1,4 @@
-#include "src/data.h"
+#include "firmware/data.h"
 #include <cstdio>
 #include <vector>
