+++ run_hls.tcl
@@ -7,7 +7,7 @@
 # open the project, don't forget to reset
 open_project -reset proj0
 set_top medium_pflow_parallel_hwopt
-add_files src/simple_pflow.cpp
+add_files firmware/simple_pflow.cpp
 add_files -tb simple_pflow_test.cpp 
 add_files -tb simple_pflow_ref.cpp
 add_files -tb DiscretePFInputs.h -cflags "-std=c++0x"
