+++ run_hls_fullpfalgo_mp7.tcl
@@ -4,7 +4,7 @@ source config_hls_fullpfalgo_mp7.tcl
 # open the project, don't forget to reset
 open_project -reset "proj3-mp7-fast"
 set_top ${l1pfTopFunc}
-add_files src/simple_fullpfalgo.cpp -cflags "-DTESTMP7"
+add_files firmware/simple_fullpfalgo.cpp -cflags "-DTESTMP7"
 add_files -tb simple_fullpfalgo_test.cpp  -cflags "-DTESTMP7 -DMP7_TOP_FUNC=${l1pfTopFunc} -DMP7_REF_FUNC=${l1pfRefFunc} -DMP7_VALIDATE=${l1pfValidate}"
 add_files -tb simple_fullpfalgo_ref.cpp -cflags "-DTESTMP7"
 add_files -tb pattern_serializer.cpp -cflags "-DTESTMP7"
@@ -21,9 +21,9 @@ set_clock_uncertainty 1.5
 config_interface -trim_dangling_port
 # do stuff
 csim_design
-#csynth_design
-#cosim_design -trace_level all
-#export_design -format ip_catalog -vendor "cern-cms" -version ${l1pfIPVersion} -description "${l1pfTopFunc}"
+csynth_design
+cosim_design -trace_level all
+export_design -format ip_catalog -vendor "cern-cms" -version ${l1pfIPVersion} -description "${l1pfTopFunc}"
 
 # exit Vivado HLS
 exit
